requirements.txt
# EthiScan - Ethical Web Vulnerability Scanner
# Python 3.10+ required

# HTTP & Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# CLI & Display
rich>=13.0.0
coloredlogs>=15.0.0

# Configuration & Validation
pydantic>=2.0.0
pyyaml>=6.0.0

# Reporting
jinja2>=3.1.0
reportlab>=4.0.0
